tests/wpt/web-platform-tests/cookies/domain/domain-attribute-missing.sub.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/cookie-helper.sub.js"></script>
</head>
<body>
  <script>
    //
    //  Set-Cookie: a=b; Path=/
    //
    test(t => {
      assert_dom_cookie("a", "b", true);
    }, "No domain attribute => available via `document.cookie`");

    async_test(t => {
      fetch("/cookies/resources/list.py", { credentials: "include" })
        .then(t.step_func(r => r.json()))
        .then(t.step_func_done(r => {
          assert_equals(r["a"], "b");
        }))
        .catch(_ => assert_unreached);
    }, "No domain attribute => sent with same-origin requests.");

    async_test(t => {
      fetch(`${SECURE_SUBDOMAIN_ORIGIN}/cookies/resources/list.py`, { credentials: "include" })
        .then(t.step_func(r => r.json()))
        .then(t.step_func_done(r => {
          assert_equals(r["a"], undefined);
        }))
        .catch(_ => assert_unreached);
    }, "No domain attribute => not sent with subdomain requests.");
  </script>
</body>
